+++ src/helpers/SiteUriHelper.php
@@ -348,11 +348,6 @@ class SiteUriHelper
         $groupedSiteUris = [];
 
         foreach ($siteUris as $siteUri) {
-            // Convert to a SiteUriModel if it is an array
-            if (is_array($siteUri)) {
-                $siteUri = new SiteUriModel($siteUri);
-            }
-
             $groupedSiteUris[$siteUri->siteId][] = $siteUri;
         }